Models online, but now with normals instead of color

+++ LaboratoryWorks/Lab_3/src/model.hpp
@@ -34,16 +34,16 @@ class Mesh{
 private:
     GLuint VAO, VBO, EBO; //vertex/element array/buffer
     std::vector <Vertex> vertices;
-    std::vector <GLuint> indices;
+    std::vector <unsigned short> indices;
     std::vector <Texture> textures;
     void prepareMesh();    
 public:
     std::vector<Vertex>& getVertices(){return vertices;};
-    std::vector<GLuint>& getIndices(){return indices;};
+    std::vector<unsigned short>& getIndices(){return indices;};
     std::vector<Texture>& getTextures(){return textures;};
     Mesh(
         std::vector<Vertex> vertices_,
-        std::vector<GLuint> indices_,
+        std::vector<unsigned short> indices_,
         std::vector<Texture> textures_
         );
     void Draw(Shader shader);
